Feat/advanced templates (#342)
* Remove obvious notes, lint, define add'l props

* Address issue #138

* Resort and refactor templates (wip)

* Update supplemental prop

* Refine (wip)

* Add header

* Define some assay constraints (wip)

* Update some assay descriptions

* Define more relationships of templates to assay

* Update test to use local config when present

* Add more HTAN mappings

* More mappings

* Revise new prop

* Update matching to assay or datatype

* Fill in descriptions, lint

* Don't run tests on draft PR

* Various ugprades (wip)

* Update data definition

* Split up templates

* Clarify and lint some assays

* More revisions

* Update base props

* Keep tests

* Revert ci

* Build jsonld

* Test with schematic v23.9.1

* Lint

* Build jsonld

* Various fixes and refinements

* Build jsonld

* Update config in tests/validate as well

* Make required

* Update validation tests

* Build jsonld

* Fix test config

* Add notes and optional asset link

* Build jsonld

* Update test expectation

* Allow auxiliary links in more places

* Build jsonld

* Remove pre-split file

* Uncomment

* Build jsonld

* Update strandedness, HTAN-style, and update tests

* Build jsonld

---------

+++ tests/generate/basic_templates.sh
@@ -24,8 +24,12 @@ else
 fi
 
 # Setup config
-echo "Getting $CONFIG to use as test config..."
-wget $CONFIG -O $TEST_CONFIG
+if [ -f "$TEST_CONFIG" ]; then
+  echo "Local $TEST_CONFIG present, running test with this local config..."
+else 
+  echo "Getting $CONFIG to use as test config..."
+  wget $CONFIG -O $TEST_CONFIG
+fi
 
 TEMPLATES=($(jq '.manifest_schemas[] | .schema_name' $TEST_CONFIG | tr -d '"'))
 #TITLES=($(jq '.manifest_schemas[] | .display_name' $TEST_CONFIG | tr -d '"'))
